cvae/requirements.txt
absl-py==0.8.1
actionlib==1.11.13
angles==1.9.11
bondpy==1.8.3
cachetools==3.1.1
camera-calibration==1.12.23
camera-calibration-parsers==1.11.13
camera-info-manager-py==0.2.3
catkin==0.7.14
catkin-pkg==0.4.14
certifi==2019.9.11
chardet==3.0.4
cv-bridge==1.12.8
cycler==0.10.0
diagnostic-analysis==1.9.3
diagnostic-common-diagnostics==1.9.3
diagnostic-updater==1.9.3
docutils==0.15.2
dynamic-reconfigure==1.5.50
gazebo-plugins==2.5.17
gazebo-ros==2.5.17
gencpp==0.6.0
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.11
gennodejs==2.0.1
genpy==0.6.7
google-auth==1.7.1
google-auth-oauthlib==0.4.1
grpcio==1.25.0
idna==2.8
image-geometry==1.12.8
interactive-markers==1.11.4
joblib==0.14.0
kiwisolver==1.1.0
laser-geometry==1.6.4
Markdown==3.1.1
matplotlib==3.0.3
message-filters==1.12.14
moveit-ros-planning-interface==0.9.15
numpy==1.17.4
oauthlib==3.1.0
openni2-launch==0.4.1
Pillow==6.2.1
pkg-resources==0.0.0
pluginlib==1.11.3
protobuf==3.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pyparsing==2.4.5
PySide2==5.13.2
python-dateutil==2.8.1
python-qt-binding==0.3.4
PyYAML==5.1.2
qt-dotgraph==0.3.11
qt-gui==0.3.11
qt-gui-cpp==0.3.11
qt-gui-py-common==0.3.11
requests==2.22.0
requests-oauthlib==1.3.0
resource-retriever==1.12.4
rosbag==1.12.14
rosboost-cfg==1.14.4
rosclean==1.14.4
roscreate==1.14.4
rosgraph==1.12.14
roslaunch==1.12.14
roslib==1.14.4
roslint==0.11.0
roslz4==1.12.14
rosmake==1.14.4
rosmaster==1.12.14
rosmsg==1.12.14
rosnode==1.12.14
rosparam==1.12.14
rospkg==1.2.0
rospy==1.12.14
rosservice==1.12.14
rostest==1.12.14
rostopic==1.12.14
rosunit==1.14.4
roswtf==1.12.14
rqt-action==0.4.9
rqt-bag==0.4.12
rqt-bag-plugins==0.4.12
rqt-console==0.4.8
rqt-dep==0.4.9
rqt-graph==0.4.9
rqt-gui==0.5.0
rqt-gui-py==0.5.0
rqt-image-view==0.4.13
rqt-launch==0.4.8
rqt-logger-level==0.4.8
rqt-moveit==0.5.7
rqt-msg==0.4.8
rqt-nav-view==0.5.7
rqt-plot==0.4.8
rqt-pose-view==0.5.8
rqt-publisher==0.4.8
rqt-py-common==0.5.0
rqt-py-console==0.4.8
rqt-reconfigure==0.4.10
rqt-robot-dashboard==0.5.7
rqt-robot-monitor==0.5.8
rqt-robot-steering==0.5.9
rqt-runtime-monitor==0.5.7
rqt-rviz==0.5.10
rqt-service-caller==0.4.8
rqt-shell==0.4.9
rqt-srv==0.4.8
rqt-tf-tree==0.5.8
rqt-top==0.4.8
rqt-topic==0.4.10
rqt-web==0.4.8
rsa==4.0
rviz==1.12.16
scikit-learn==0.21.3
scipy==1.3.3
sensor-msgs==1.12.7
shiboken2==5.13.2
six==1.13.0
sklearn==0.0
smach==2.0.1
smach-ros==2.0.1
smclib==1.8.3
srdfdom==0.4.2
tensorboard==2.0.1
tensorboardX==1.9
tf==1.11.9
tf-conversions==1.11.9
tf2-geometry-msgs==0.5.19
tf2-kdl==0.5.19
tf2-py==0.5.19
tf2-ros==0.5.19
topic-tools==1.12.14
torch==1.2.0+cu92
torchvision==0.4.0+cu92
trac-ik-python==1.5.0
urdfdom-py==0.3.3
urllib3==1.25.7
Werkzeug==0.16.0
xacro==1.11.3
